dont log visits for guests

+++ core_dev/core/functions_visits.php
@@ -11,7 +11,7 @@
 	function logVisit($_type, $_owner)
 	{
 		global $db, $session;
-		if (!is_numeric($_type) || !is_numeric($_owner)) return false;
+		if (!$session->id || !is_numeric($_type) || !is_numeric($_owner)) return false;
 
 		//only log the latest entry for each visitor
 		$q = 'DELETE FROM tblVisits WHERE type='.$_type.' AND ownerId='.$_owner.' AND creatorId='.$session->id;
